+++ packages/common/src/lib/common/constants.ts
@@ -24,6 +24,7 @@ const SUPPORTED_FEATURES_INTERNAL = [
     'CONTROLLER_CMD_EXEC_FILES', // cmdExec host message supports sending files together with the command. Since `js-controller` 7.2
     'CONTROLLER_FEATURE_REQUEST', // js-controller supports feature support requests via host messages. Since `js-controller` 7.2
     'CONTROLLER_USED_RESOURCES', // js-controller maintains a registry of exclusive resources (serial/TCP/UDP ports, ...) used by instances. Since `js-controller` 8.0
+    'CONTROLLER_PYTHON_ADAPTERS', // Controller can start adapters with `common.platform: "Python"`. Since `js-controller` 8.0
 ] as const;
 
 export const SUPPORTED_FEATURES = [...SUPPORTED_FEATURES_INTERNAL];